Change how public key is supplied
Key pair is not created on-the-fly anymore, but user provides the public
key as an existing file in a variable. It's possible to use the standard
~/.ssh/id_rsa.pub. Once all resources are created, it's then possible to
log in to the instances with `ssh -i ~/.ssh/id_rsa ubuntu@<IP>`.

If the variable value is ever changed (e.g. to `~/.ssh/other.pub` and a
Terraform apply is run, then Terraform replaces the AWS Key Pair and
also the instances. The instances will be set up with the new key pair,
thus logging in has to be done with `ssh -i ~/.ssh/other ubuntu@<IP>`.